# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Ruta - API de verificacion de entregas",
    "description" : "Verificacion de entregas con PIN, foto y ubicacion. Todas las rutas /api/v1/admin/** y /api/v1/driver/** exigen la cookie HttpOnly de sesion (\"access_token\") que POST /api/v1/auth/login establece via Set-Cookie. Las peticiones que modifican datos requieren ademas el header X-XSRF-TOKEN (proteccion CSRF con doble cookie). Swagger UI envia la cookie solo si el login se hizo desde el propio navegador con \"Try it out\".",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Mismo origen que la aplicacion"
  } ],
  "security" : [ {
    "sessionCookie" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/admin/invoices" : {
      "get" : {
        "tags" : [ "Facturas (admin)" ],
        "summary" : "Listar facturas",
        "description" : "Facturas filtradas por numero o cliente, paginado (tamano maximo 100), incluido su estado y PIN vigente.",
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : ""
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "Tamano de pagina (maximo 100).",
          "required" : false,
          "schema" : {
            "maximum" : 100,
            "type" : "integer",
            "description" : "Tamano de pagina (maximo 100).",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "'size' supera el maximo permitido",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "Pagina de facturas",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseAdminInvoiceResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Facturas (admin)" ],
        "summary" : "Crear una factura",
        "description" : "Crea la factura en borrador, con sus lineas de producto. No genera PIN todavia: eso ocurre al publicarla.",
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateInvoice"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Datos invalidos (campos requeridos, coordenadas fuera de rango, sin productos)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Ya existe una factura con ese numero",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201" : {
            "description" : "Factura creada; el header Location apunta al recurso",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminInvoiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/invoices/{id}/publish" : {
      "post" : {
        "tags" : [ "Facturas (admin)" ],
        "summary" : "Publicar una factura",
        "description" : "Genera el PIN de 6 digitos y deja la factura lista para que el conductor la confirme. El PIN se debe comunicar al cliente por un canal operativo aparte (Fase1 §3).",
        "operationId" : "publish",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Factura publicada, con su PIN",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminInvoiceResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "La factura no existe o ya estaba publicada",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreateInvoice" : {
        "required" : [ "number", "partnerName", "products" ],
        "type" : "object",
        "properties" : {
          "number" : {
            "maxLength" : 60,
            "minLength" : 0,
            "type" : "string",
            "example" : "001-104-0000001234"
          },
          "partnerName" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string",
            "example" : "Cliente de prueba"
          },
          "deliveryAddress" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string",
            "example" : "Calle de prueba, Guayaquil"
          },
          "latitude" : {
            "maximum" : 90,
            "exclusiveMaximum" : false,
            "minimum" : -90,
            "exclusiveMinimum" : false,
            "type" : "number",
            "format" : "double",
            "example" : -2.170998
          },
          "longitude" : {
            "maximum" : 180,
            "exclusiveMaximum" : false,
            "minimum" : -180,
            "exclusiveMinimum" : false,
            "type" : "number",
            "format" : "double",
            "example" : -79.922359
          },
          "requiresPin" : {
            "type" : "boolean",
            "description" : "Si es true, publicar() genera un PIN de 6 digitos.",
            "default" : true
          },
          "products" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Product"
            }
          }
        }
      },
      "Product" : {
        "required" : [ "description", "quantity" ],
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string",
            "example" : "Arroz blanco - funda 1 kg"
          },
          "quantity" : {
            "type" : "number",
            "format" : "double",
            "example" : 2
          }
        }
      },
      "AdminInvoiceResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "number" : {
            "type" : "string"
          },
          "partnerName" : {
            "type" : "string"
          },
          "deliveryAddress" : {
            "type" : "string"
          },
          "invoiceDate" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string",
            "enum" : [ "draft", "posted", "cancel" ]
          },
          "requiresPin" : {
            "type" : "boolean"
          },
          "pin" : {
            "type" : "string"
          },
          "confirmed" : {
            "type" : "boolean"
          },
          "expectedLatitude" : {
            "type" : "number",
            "format" : "double"
          },
          "expectedLongitude" : {
            "type" : "number",
            "format" : "double"
          },
          "createdBy" : {
            "type" : "string"
          },
          "publishedBy" : {
            "type" : "string"
          }
        }
      },
...
      "PageResponseAdminInvoiceResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AdminInvoiceResponse"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "description" : "Mensaje legible para mostrar al usuario.",
            "example" : "El PIN ingresado no es correcto."
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string",
              "description" : "Detalle por campo, solo presente en errores de validacion (400 de Bean Validation).",
              "example" : "{\"pin\":\"El PIN debe tener 6 digitos\"}"
            },
            "description" : "Detalle por campo, solo presente en errores de validacion (400 de Bean Validation).",
            "example" : {
              "pin" : "El PIN debe tener 6 digitos"
            }
          },
          "path" : {
            "type" : "string",
            "description" : "Ruta donde ocurrio el error.",
            "example" : "/api/v1/driver/deliveries/confirm"
          },
          "timestamp" : {
            "type" : "string",
            "description" : "Instante del error, en UTC.",
            "format" : "date-time"
          }
        },
        "description" : "Forma unica de error de la API. Todas las respuestas 4xx/5xx usan este contrato."
      }
    },
    "securitySchemes" : {
      "sessionCookie" : {
        "type" : "apiKey",
        "name" : "access_token",
        "in" : "cookie"
      }
    }
  }
}